+++ R/ume.network.R
@@ -271,7 +271,9 @@ ume.binomial.rjags <- function(network){
                    "\n\t\tdelta[i,1] <- 0",
                    "\n\t\tmu[i] ~ dnorm(mean.mu,prec.mu)",
                    "\n\t\tfor(k in 1:na[i]) {",
-                   "\n\t\t\tr[i,k] ~ dbin(p[i,k], n[i,k])")
+                   "\n\t\t\tr[i,k] ~ dbin(p[i,k], n[i,k])",
+                   "\n\t\t\tlogit(p[i,k]) <- mu[i] + delta[i,k]"
+                   )
     
     if(type == "fixed"){
       code <- paste0(code, "\n\t\t\tlogit(p[i,k]) <- mu[i] + d[t[i,1], t[i,k]]")
@@ -289,6 +291,10 @@ ume.binomial.rjags <- function(network){
     if(type == "random"){
       code <- paste0(code, "\n\t\tfor (k in 2:na[i]) {",
                      "\n\t\t\tdelta[i,k] ~ dnorm(d[t[i,1],t[i,k]], prec)",
+                     "\n\t\t}")
+    } else if(type == "fixed"){
+      code <- paste0(code, "\n\t\tfor (k in 2:na[i]) {",
+                     "\n\t\t\tdelta[i,k] <- d[t[i,1],t[i,k]]",
                      "\n\t\t}")
     }